requirements.txt
-f /usr/share/pip-wheels
amqp==1.4.9
anyjson==0.3.3
Babel==2.5.3
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
diff-match-patch==20121119
Django==1.11.10
django-appconf==1.0.2
django-celery==3.2.2
django-compressor==2.2
django-filter==0.5.3
django-formtools==2.1
django-good-otp==1.1.2
django-import-export==1.0.0
django-ipware==2.0.2
django-otp==0.4.3
django-pagination==1.0.7
django-phonenumber-field==1.3.0
django-twilio==0.9.0
django-two-factor-auth==1.7.0
et-xmlfile==1.0.1
idna==2.6
image==1.5.20
jdcal==1.4
Jinja2==2.10
kombu==3.0.37
MarkupSafe==1.0
mysqlclient==1.3.12
odfpy==1.3.6
openpyxl==2.5.3
phonenumberslite==8.9.2
Pillow==5.0.0
pycparser==2.18
PyJWT==1.6.1
pyotp==2.2.6
PySocks==1.6.8
python-http-client==3.0.0
pytz==2018.3
PyYAML==3.12
qrcode==4.0.4
rabbitmq==0.2.0
rcssmin==1.0.6
requests==2.18.4
rjsmin==1.0.12
sendgrid==5.3.0
six==1.11.0
tablib==0.12.1
twilio==6.11.0
unicodecsv==0.14.1
urllib3==1.22
vine==1.1.4
xlrd==1.1.0
xlwt==1.3.0
djangorestframework==3.8.2